requirements.txt
Flask==2.0.3
Flask-MySQL==1.5.2
Jinja2==3.0.3
gunicorn==20.1.0
pyyaml==6.0
azure-storage-blob==12.10.0
python-dotenv==0.19.2